Backend/requirements-dev.txt
# =====================================================
# DEPENDENCIAS DE DESARROLLO - Sistema Médico Integral
# Versión: Sprint 2 (Diciembre 2025)
# =====================================================

# Incluir dependencias base
-r requirements-base.txt

# Testing
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-asyncio==0.23.8

# Linting y Formateo
ruff==0.8.4
mypy==1.14.0

# Type stubs
types-requests==2.31.0.20241016
types-python-dateutil==2.9.0.20241206

# Debugging
ipdb==0.13.13
